+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2017-10-02
+### Added
+- main parseNext function
+- added public functions (parse and now function) 
+
 ## [0.1.1] - 2017-09-30
 ### Update
 - README update